my life

+++ 0x0C-more_malloc_free/2-calloc.c
@@ -7,7 +7,7 @@
  * 
  */
 void *_calloc(unsigned int nmemb, unsigned int size)
-{int i;
+{size_t i;
 if(nmemb==0||size==0)
 {
 return(NULL);
